Document usage of master container registry tag
Quay is supposed to tag images as latest if it's building from the
default branch, i.e. master. This is currently not happening when a
custom git build trigger is used. A PR has been filed to fix this
behavior. If/when approved and deployed, we should revert this commit.

Quay is currently closed source so not all will have access to its
private repo:
    https://github.com/quay/quay/pull/3530

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>