Show all active/inactive flags in a request
Currently, if a flag is active when a request is made and then it is made inacitve, it will not show up in the response when the user requests to see his/her past request. This change will show that flag in the response even though it is inactive now.