Refactor the way requirements are defined in setup.py
Since to run the API you must have the dependencies of the workers
installed in order to call the Celery tasks, the workers requirements
should not be extra. They are always required no matter the deployment.

Now, there is an extras_require for `web` which just contains the
dependencies for the API.