Add go-package without dependencies to request
This patch fixes the case where a go package without dependencies would
be processed. Before, the go package would not be added to the request.
This patch ensures the package is added to the request by explicitly
calling the method to add packages to a request instead of relying on
the logic of adding dependencies to a package.

Signed-off-by: Athos Ribeiro <athos@redhat.com>