Add groovy script to clean pip volatile resources
After a request is moved to the stale state, we should cleanup the
temporary Nexus repositories, user, and role created for the request.
This requires a new nexus script to be registered. This patch introduces
such script.

Note that the current script makes Nexus raise exceptions when removing
resources. This is a known issue, already present in the JS analogous
script. The exceptions are safe to ignore and should be fixed in Nexus
3.22. See https://issues.sonatype.org/browse/NEXUS-22829 for further
information.

Signed-off-by: Athos Ribeiro <athos@redhat.com>