Handle pip duplicated dependencies
If the same pip dependency is listed in different packages for a same
request, Cachito will try to push that dependency to the Nexus temporary
request repository twice. If that happens, an exception will be raised.
This patch catches this exception and makes sure the component is
already present in the temporary repository. When this is the case,
Cachito will proceed with the request.

Signed-off-by: Athos Ribeiro <athos@redhat.com>