Simplify how replacements are applied
The code does not need to check if nexus replacements are empty or not.
The function that replaces package.json already returns None if no
replacements apply. For yarn.lock, the only time we would want None is
if there are no dependencies with resolved urls, i.e. only 'file:'
dependencies or no dependencies.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>