Fix how JS code searches for Nexus components
CLOUDBLD-4938

The Nexus search API does not allow the user to search specifically for
components without groups. If no group is specified, Nexus returns all
the matching components regardless of group.

In some cases, this causes Cachito to fail. For example, if a Yarn
package requires both `some-package` and `@types/some-package` (at the
same version), Cachito will eventually search for `some-package` and
fail because 2 different results are found.

Implement an ad-hoc workaround where the search result is filtered by
None afterwards. Keep the default behaviour the same as before (match
regardless of group).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>